Move some conversions from `utils` to DAL
